# perf: faster utf8<->utf16 conversion on Windows