@aras-p 2024-12-01
The one CI failure fails at cmake setup time with `Could NOT find pystring (missing: pystring_LIBRARY) (found suitable version "1.1.4", minimum required is "1.1.4")` which is... confusing.