@lgritz 2024-12-01
I have seen that from time to time recently... but not always, since we obviously do have passing tests most of the time. I'm not quite sure what's going on there, but it's obviously unrelated to your patch. I will poke it to make it run the CI again for that job, but even if that doesn't pass, I won't hold this up.

OIIO itself doesn't use pystring, that's being pulled in by the OpenColorIO build.